moved Event.kt over to new model folder
