+++ app/src/main/java/com/example/nsc_events/model/Event.kt
@@ -1,4 +1,4 @@
-package com.example.nsc_events.data
+package com.example.nsc_events.model
 
 import java.util.Date
 
@@ -39,4 +39,3 @@ data class Event(
 
     }
 }
-
