+++ app/src/main/java/com/example/nsc_events/screen/EventDetail.kt
@@ -7,7 +7,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.example.nsc_events.data.Event
+import com.example.nsc_events.model.Event
 
 @Composable
 fun DeleteButton(
